% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPxplorer-package.R
\docType{package}
\name{CPxplorer-package}
\alias{CPxplorer}
\alias{CPxplorer-package}
\title{CPxplorer: Interactive tools for PCAs and analogues analysis and quantification}
\description{
Generates list of ions of selected PCAs and analogues for mass spectrometric analysis and perform quantification by deconvolution.
}
\author{
\strong{Maintainer}: Thanh Wang \email{thanh.wang@liu.se} (\href{https://orcid.org/0000-0002-5729-1908}{ORCID})

Authors:
\itemize{
  \item Idoia Beloqui Ezquer \email{idoia.beloqui.ezquer@liu.se}
}

}
\keyword{internal}
